[BACKLOG-6739] Relocated "will/did/rejected" event classes.
