+++ package-res/resources/web/pentaho/lang/events/RejectedChange.js
@@ -14,33 +14,30 @@
  * limitations under the License.
  */
 define([
-  "pentaho/lang/Event",
-  "pentaho/util/error"
+  "../Event",
+  "../../util/error"
 ], function(Event, utilError) {
   "use strict";
 
-  return Event.extend("pentaho.visual.base.events.DidChange",
-    /** @lends pentaho.visual.base.events.DidChange# */{
-      constructor: function(source, value, will) {
+  return Event.extend("pentaho.visual.base.events.RejectChange",
+    /** @lends pentaho.visual.base.events.RejectChange# */{
+      constructor: function(source, error, will) {
+        if(!error) throw utilError.argRequired("error");
         if(!will) throw utilError.argRequired("will");
 
-        this.base("did:change", source, false);
-        this._property = will.property;
-        this._value = value;
+        this.base("rejected:change", source, false);
+        this._error = error;
         this._previousValue = will.previousValue;
-      },
-      get property(){
-        return this._property;
-      },
-      get value(){
-        return this._value;
       },
       get previousValue(){
         return this._previousValue;
+      },
+      get error() {
+        return this._error;
       }
     },{
       get type() {
-        return "did:change";
+        return "rejected:change";
       }
     });
